Refactor processing module into submodules
