+++ backend/src/processing/mod.rs
@@ -1,0 +1,4 @@
+pub mod ocr;
+pub mod ai;
+pub mod parse;
+pub mod report;
